requirements.txt
aio-pika==9.1.4
aiormq==6.7.6
aiosmtplib==2.0.2
anyio==3.7.0
arrow==1.2.3
attrs==23.1.0
automata-lib==7.1.0
binaryornot==0.4.4
blinker==1.6.2
build==0.10.0
CacheControl==0.12.14
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
cookiecutter==1.7.3
crashtest==0.4.1
cryptography==37.0.4
distlib==0.3.6
dnspython==2.3.0
dulwich==0.21.5
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.97.0
fastapi-mail==1.3.1
filelock==3.12.2
frozendict==2.3.8
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httptools==0.5.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==6.0.0
installer==0.7.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
manage-fastapi==1.1.1
MarkupSafe==2.1.3
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
packaging==23.1
pamqp==3.2.1
pexpect==4.8.0
pika==1.3.2
pipenv==2023.6.18
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs==3.8.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
poyo==0.5.0
prometheus-client==0.17.0
prometheus-fastapi-instrumentator==6.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.9
pydot==1.4.2
pyparsing==3.1.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
PyYAML==6.0
questionary==1.10.0
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.5.24
typer==0.4.2
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
watchfiles==0.19.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
yarl==1.9.2
zipp==3.15.0
